backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
Pillow = "^9.0.1"
psycopg2-binary = "^2.9.3"
django-cors = "^0.1"
django-cors-headers = "^3.11.0"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
